feat: やっぱりApp ディレクトリはやめる方針に変更
