--- src/app/register/page.tsx
@@ -1,9 +1,0 @@
-'use client'
-
-import { NextPage } from 'next'
-
-const register: NextPage = () => {
-  return <>登録ページ</>
-}
-
-export default register
