+++ src/pages/_app.tsx
@@ -1,17 +1,12 @@
-'use client'
-
 import { createTheme, ThemeProvider } from '@mui/material'
 import axios, { AxiosError } from 'axios'
 import { memo, useEffect } from 'react'
 import { RecoilRoot, useSetRecoilState } from 'recoil'
 import useSWR, { SWRConfig } from 'swr'
-import './globals.scss'
+import '@/styles/globals.scss'
 import { authUserState, natureDataState, pokemonDataState } from '@/store'
-
-// export const metadata = {
-//   title: 'Create Next App',
-//   description: 'Generated by create next app',
-// }
+import { AppProps } from 'next/app'
+import { Header } from '@/components/organisms/Header'
 
 const AppInit = memo(() => {
   const setAuthUser = useSetRecoilState(authUserState)
@@ -30,7 +25,7 @@ const AppInit = memo(() => {
   return <></>
 })
 
-const RootLayout = ({ children }: { children: React.ReactNode }) => {
+export default function App({ Component, pageProps }: AppProps) {
   axios.defaults.baseURL = process.env.NEXT_PUBLIC_BASE_URL
   axios.defaults.withCredentials = true
 
@@ -59,18 +54,16 @@ const RootLayout = ({ children }: { children: React.ReactNode }) => {
   }
 
   return (
-    <html lang="ja">
-      <ThemeProvider theme={theme}>
-        <RecoilRoot>
-          <SWRConfig value={swrConfigValue}>
-            {/* <Header /> */}
-            <AppInit />
-            <body>{children}</body>
-          </SWRConfig>
-        </RecoilRoot>
-      </ThemeProvider>
-    </html>
+    <ThemeProvider theme={theme}>
+      <RecoilRoot>
+        <SWRConfig value={swrConfigValue}>
+          <Header />
+          <AppInit />
+          <main>
+            <Component {...pageProps} />
+          </main>
+        </SWRConfig>
+      </RecoilRoot>
+    </ThemeProvider>
   )
 }
-
-export default RootLayout
