+++ src/pages/register/index.tsx
@@ -1,0 +1,3 @@
+export default function Register() {
+  return <>登録ページ</>
+}
